chore(orm): restructure orm
create orm folder
abstract has_many and belongs_to into model_associations.rb
transfer helpers/model_helpers into orm folder
update ui for sample app

[Finishes #122063883]
