+++ spec/todolist/app/views/items/new.erb
@@ -3,12 +3,12 @@
     <h5>Create New Item</h5>
     <form action="/items" name="items" method="post">
       <div class="input-field">
-        <input name="items[name]" id="item-name" class="validate">
         <label class="active" id="item-name">Item Name</label>
+        <input name="items[name]" id="item-name" class="validate">
       </div>
       <div class="input-field">
-        <input name="items[status]" id="item-status" class="validate">
         <label class="active" id="item-name">Status</label>
+        <input name="items[status]" id="item-status" class="validate">
       </div>
       <input type="submit" class="btn" id="item-submit"/>
     </form>
