+++ spec/todolist/app/views/layouts/application.erb
@@ -4,6 +4,7 @@
       <link href="http://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
       <link type="text/css" rel="stylesheet" href="/css/materialize.min.css"  media="screen,projection"/>
       <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+      <title>TodoApp With Rolletskates</title>
     </head>
 
     <body>
