+++ spec/todolist/app/views/posts/edit.erb
@@ -4,10 +4,16 @@
       <h1>Edit Post</h2>
 
       <form action="/posts/<%= @post.id %>" method="post">
-        <input type="hidden" name="_method" value="put"><br>
-        <input type="text" name="posts[title]" value="<%= @post.title %>"><br>
-        <textarea name="posts[description]"><%= @post.description %></textarea><br>
-        <input type="submit" value="Update"/>
+        <input type="hidden" name="_method" value="put">
+        <div class="input-field">
+          <label class="active">Post Title</label>
+          <input type="text" name="posts[title]" value="<%= @post.title %>"><br>
+        </div>
+        <div class="input-field">
+          <label class="active">Post Body</label>
+          <textarea name="posts[description]" class="materialize-textarea"><%= @post.description %></textarea>
+        </div>
+        <input type="submit" class="btn" value="Update"/>
       </form>
     </div>
   </div>
