+++ spec/todolist/app/views/posts/show.erb
@@ -2,28 +2,45 @@
 <div class="container">
   <div class="row">
     <div class="col l8 offset-l2">
-      <h3><%= @post.title %></h3>
+      <h3><%= @post.title %>
+        <span class="right valign-center">
+          <a href="/posts/<%= @post.id %>/edit"><i class="material-icons">edit</i></a>
+          <form action="/posts/<%= @post.id %>" method="post" style="display: inline"><input type="hidden" name="_method" value="delete"/><button type="submit" style="background-color: rgba(255,255,255, 0); border: 0px; padding: 0px;"><i class="material-icons">delete</i></button></form>
+        </span>
+      </h3>
       <div class="flow-text"><%= @post.description %></div>
-      <h6>Comments</h6>
-      <% require "pry"; binding.pry %>
-      <hr>
-      <ul>
-        <% @post.comments.each do |comment| %>
-          <li>
-            <%= comment.name %> <br>
-            <%= comment.message %> <br>
-            <a href="/comments/<%= comment.id %>/edit">EDIT</a> | <form action="/comments/<%= comment.id %>" method="comment" style="display: inline"><input type="hidden" name="_method" value="delete"/><input type="submit" value="DELETE"/></form></li>
-        <% end %>
-      </ul>
+<hr />
+      <div class="row">
+        <h6>Comments</h6>
+        <div class="col l12">
+          <% @post.comments.each do |comment| %>
+            <div class="card-panel teal">
+              <div class="right"><form action="/comments/<%= comment.id %>" method="post" style="display: inline"><input type="hidden" name="_method" value="delete"/><button type="submit" style="background-color: rgba(255,255,255, 0); border: 0px; padding: 0px;"><i class="material-icons">delete</i></button></form></div>
+              <div class="white-text"><%= comment.message %></div>
+              <div class="white-text right">- <%= comment.name %></div>
+            </div>
+          <% end %>
+        </div>
+      </div>
 
-      <h5>New Comment</h5>
+    <div class="container">
+      <div class="row">
+        <h5>New Comment</h5>
+          <form action="/comments" name="comments" method="post">
+            <div class="input-field">
+              <label class="active">Comment</label>
+              <textarea name="comments[message]" class="materialize-textarea"></textarea>
+            </div>
+            <div class="input-field">
+              <label class="active">Name</label>
+              <input type="text" name="comments[name]">
+            </div>
+            <input type="hidden" name="comments[post_id]" value="<%= @post.id %>"/><br>
 
-      <form action="/comments" name="comments" method="post">
-        <input type="text" name="comments[name]"><br>
-        <textarea name="comments[message]"></textarea><br>
-        <input type="hidden" name="comments[post_id]" value="<%= @post.id %>"/><br>
-        <input type="submit"/>
-      </form>
+            <input type="submit" class="btn"/>
+          </form>
+        </div>
+      </div>
     </div>
   </div>
 </div>
